archive/examples/python/requirements.txt
# PAW Blockchain Python SDK Requirements

# Cosmos SDK Python libraries
cosmpy==0.9.0
ecdsa==0.18.0
bech32==1.2.0
mnemonic==0.20

# Crypto libraries
cryptography==41.0.5
pycryptodome==3.19.0

# HTTP and networking
requests==2.31.0
websockets==12.0
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
